fixed all of mine except swirl. blend might be broken hope not.

+++ imageManip.h
@@ -8,7 +8,7 @@
 
 //not done need 2 more functions
 Image * exposure(Image * img, int ev);
-Image * blend(Image *img, Image *img2, int alpha);
+Image * blend(Image *img, Image *img2, float alpha);
 Image * zoom_in(Image *img);
 Image * zoom_out(Image *img);
 Image * swirl(Image *img, int xCenter, int yCenter, int distortion);
